Update lambda_function.py
if diff_dates(strUpdate, now) < int(intHours):

This line breaks things, as intHours is in seconds and strUpdate and now are both in epoch time. Even if intHours was in the correct format all this does is make old alerts occur if they're over 24 hours old.

You need a try statement around isItemResponse = response.get('Item'), as if the item isn't in the database it'll throw an error because response.get is invalid for NoneType.